Remove the dependency on spacevim
All that I was using from spacevim was the keybindings. It made sense
to just put that part of the plugin directly into the rc. The MIT
license for spacevim is included in conjunction with the spacevim code
and applies only to it.

This also removes the dependency on swoop.